Reduce frontend copy pasta
