+++ cmd/management-console/web/src/ProductCertificate.scss
@@ -1,17 +1,6 @@
 .product-certificate {
     &__bg {
-        position: absolute;
-        user-select: none;
-        pointer-events: none;
-        z-index: 0;
-        top: 0;
-        bottom: 0;
-        left: 0;
-        right: 0;
-        background-image: url(./bg-sprinkles-2x.png);
         background-color: #151c28;
-        background-repeat: none;
-        background-size: cover;
         opacity: 0.35;
         background-blend-mode: overlay;
     }
@@ -28,5 +17,11 @@
         z-index: 9;
         background-color: rgba(43, 55, 80, 0.85);
         font-size: 0.875em;
+    }
+}
+
+.product-certificate-parcel {
+    .product-certificate__bg {
+        background-image: url(./bg-sprinkles-2x.png);
     }
 }
