+++ shared/src/index.scss
@@ -16,4 +16,6 @@
 @import './panel/Panel';
 @import './panel/views/FileLocations';
 @import './panel/views/HierarchicalLocationsView';
+@import './productSubscription/ProductCertificate';
+@import './productSubscription/TrueUpStatusSummary';
 @import './symbols/SymbolIcon';
