+++ shared/src/productSubscription/ProductCertificate.scss
@@ -1,0 +1,45 @@
+.product-certificate {
+    &__bg {
+        position: absolute;
+        user-select: none;
+        pointer-events: none;
+        z-index: 0;
+        top: 0;
+        bottom: 0;
+        left: 0;
+        right: 0;
+        background-repeat: none;
+        background-size: cover;
+    }
+
+    &__logo {
+        height: 5rem;
+        z-index: 9;
+    }
+    .btn {
+        z-index: 9;
+    }
+    .card-footer {
+        z-index: 9;
+    }
+}
+
+.theme-light {
+    .product-certificate {
+        &__bg {
+            opacity: 0.125;
+        }
+    }
+}
+
+.theme-dark {
+    .product-certificate {
+        &__bg {
+            opacity: 0.35;
+            background-blend-mode: overlay;
+        }
+        .card-footer {
+            background-color: rgba(43, 55, 80, 0.85);
+        }
+    }
+}
