+++ web/src/enterprise.scss
@@ -7,7 +7,6 @@
 @import './enterprise/extensions/registry/RegistryNewExtensionPage';
 @import './enterprise/extensions/extension/RegistryExtensionManagePage';
 @import './enterprise/productSubscription/ProductCertificate';
-@import './enterprise/productSubscription/TrueUpStatusSummary';
 @import './enterprise/user/productSubscriptions/PaymentTokenFormControl';
 
 @import './enterprise/extensions/explore/ExtensionsExploreSection';
