+++ web/src/enterprise/productSubscription/ProductCertificate.scss
@@ -1,17 +1,6 @@
 .product-certificate {
     &__bg {
-        position: absolute;
-        user-select: none;
-        pointer-events: none;
-        z-index: 0;
-        top: 0;
-        bottom: 0;
-        left: 0;
-        right: 0;
-        background-image: url(/.assets/img/bg-sprinkles-2x.png);
         background-color: var(--body-bg);
-        background-repeat: none;
-        background-size: cover;
     }
 
     &__logo {
@@ -27,21 +16,17 @@
 }
 
 .theme-light {
-    .product-certificate {
-        &__bg {
-            opacity: 0.125;
+    .product-certificate-webpack {
+        .product-certificate__bg {
+            background-image: url(/.assets/img/bg-sprinkles-2x.png);
         }
     }
 }
 
 .theme-dark {
-    .product-certificate {
-        &__bg {
-            opacity: 0.35;
-            background-blend-mode: overlay;
-        }
-        .card-footer {
-            background-color: rgba(43, 55, 80, 0.85);
+    .product-certificate-webpack {
+        .product-certificate__bg {
+            background-image: url(/.assets/img/bg-sprinkles-2x.png);
         }
     }
 }
