--- web/src/enterprise/productSubscription/TrueUpStatusSummary.scss
@@ -1,15 +1,0 @@
-.true-up-status-summary {
-    &__container {
-        display: flex;
-        justify-content: space-between;
-        flex-wrap: wrap;
-    }
-
-    &__item {
-        width: 32%;
-        @media (max-width: $media-md) {
-            width: 100%;
-            margin-bottom: 1rem;
-        }
-    }
-}
